moved and registered the config option
